etl_pipeline.py
"""
Healthcare Claims ETL Pipeline
A minimal placeholder/demo for CI/CD pipeline showcase
"""

import os
from typing import List, Dict, Any
from dotenv import load_dotenv

# Load environment variables
load_dotenv()


class HealthcareETL:
    """ETL pipeline for processing healthcare claims data"""

    def __init__(self):
        """Initialize ETL pipeline with database configuration"""
        self.db_config = {
            'host': os.getenv('DB_HOST', 'localhost'),
            'database': os.getenv('DB_NAME', 'healthcare_db'),
            'user': os.getenv('DB_USER', 'user'),
            'password': os.getenv('DB_PASSWORD', 'password')
        }

    def extract_claims(self) -> List[Dict[str, Any]]:
        """
        Extract healthcare claims from source database
        
        Returns:
            List of claim dictionaries
        """
        # Placeholder: In real implementation, this would query a database
        sample_claims = [
            {
                'claim_id': 'CLM001',
                'patient_id': 'PAT001',
                'provider_id': 'PROV001',
                'service_date': '2024-01-15',
                'procedure_code': '99213',
                'amount': 150.00,
                'status': 'pending'
            },
            {
                'claim_id': 'CLM002',
                'patient_id': 'PAT002',
                'provider_id': 'PROV001',
                'service_date': '2024-01-16',
                'procedure_code': '99214',
                'amount': 200.00,
                'status': 'approved'
            },
            {
                'claim_id': 'CLM003',
                'patient_id': 'PAT001',
                'provider_id': 'PROV002',
                'service_date': '2024-01-17',
                'procedure_code': '99215',
                'amount': 250.00,
                'status': 'pending'
            }
        ]
        return sample_claims

    def transform_claims(self, claims: List[Dict[str, Any]]) -> List[Dict[str, Any]]:
        """
        Transform and validate healthcare claims
        
        Args:
            claims: List of raw claim dictionaries
            
        Returns:
            List of transformed claim dictionaries
        """
        transformed_claims = []
        
        for claim in claims:
            # Apply business logic transformations
            transformed_claim = {
                'claim_id': claim['claim_id'],
                'patient_id': claim['patient_id'],
                'provider_id': claim['provider_id'],
                'service_date': claim['service_date'],
                'procedure_code': claim['procedure_code'],
                'amount': claim['amount'],
                'status': claim['status'].upper(),
                'processed_date': '2024-01-20',
                'is_valid': self._validate_claim(claim)
            }
            transformed_claims.append(transformed_claim)
        
        return transformed_claims

    def _validate_claim(self, claim: Dict[str, Any]) -> bool:
        """
        Validate a single claim
        
        Args:
            claim: Claim dictionary to validate
            
        Returns:
            True if valid, False otherwise
        """
        # Basic validation rules
        if not claim.get('claim_id'):
            return False
        if not claim.get('procedure_code'):
            return False
        if claim.get('amount', 0) <= 0:
            return False
        return True

    def load_claims(self, claims: List[Dict[str, Any]]) -> bool:
        """
        Load transformed claims into target database
        
        Args:
            claims: List of transformed claim dictionaries
            
        Returns:
            True if load successful, False otherwise
        """
        # Placeholder: In real implementation, this would insert into Azure SQL
        try:
            print(f"Loading {len(claims)} claims to database...")
            for claim in claims:
                print(f"  - Claim {claim['claim_id']}: {claim['status']}")
            return True
        except Exception as e:
            print(f"Error loading claims: {e}")
            return False

    def run_pipeline(self) -> Dict[str, Any]:
        """
        Execute the complete ETL pipeline
        
        Returns:
            Dictionary with pipeline execution results
        """
        results = {
            'extracted': 0,
            'transformed': 0,
            'loaded': 0,
            'success': False
        }
        
        try:
            # Extract
            claims = self.extract_claims()
            results['extracted'] = len(claims)
            
            # Transform
            transformed_claims = self.transform_claims(claims)
            results['transformed'] = len(transformed_claims)
            
            # Load
            load_success = self.load_claims(transformed_claims)
            if load_success:
                results['loaded'] = len(transformed_claims)
                results['success'] = True
            
        except Exception as e:
            print(f"Pipeline error: {e}")
        
        return results


def main():
    """Main entry point for the ETL pipeline"""
    etl = HealthcareETL()
    results = etl.run_pipeline()
    
    print("\n=== Pipeline Results ===")
    print(f"Extracted: {results['extracted']} claims")
    print(f"Transformed: {results['transformed']} claims")
    print(f"Loaded: {results['loaded']} claims")
    print(f"Success: {results['success']}")
    
    return results


if __name__ == "__main__":
    main()
